+++ pysui/sui/sui_txn/async_transaction.py
@@ -246,7 +246,13 @@ class SuiTransactionAsync(_SuiTransactionBase):
     @versionadded(
         version="0.17.0", reason="Convenience for serializing and dry-running."
     )
-    async def get_transaction_data(self, *, gas_budget) -> bcs.TransactionData:
+    @versionchanged(
+        version="0.35.1",
+        reason="gas_budget is optional.",
+    )
+    async def get_transaction_data(
+        self, *, gas_budget: Optional[str] = None
+    ) -> bcs.TransactionData:
         """Returns the BCS TransactionKind."""
         return await self._build_for_execute(gas_budget)